feat(protocol): Improve protocol generator and make it more flexible.
